Cleanup and improvements
